requirements.txt
tinkoff-investments>=0.2.0b59
redis>=4.6.0
python-dotenv>=1.0.0
# ib_insync>=0.9.86
aiogram>=2.25.1
